refactor: 잘못 추가된 패티지 경로 수정
