+++ presentation/src/main/java/com/yapp/presentation/ui/createproject/CreateProjectActivity.kt
@@ -3,6 +3,7 @@ package com.yapp.presentation.ui.createproject
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import com.yapp.presentation.ui.createproject.screen.CreateProjectScreen
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
